[WIP/RFC] create a proper README.md, both for users and developers
Now that the addon has a more visible README we might as well use it. This may help attract contributors, especially if we expand on the section that explains in details how the addons is structured code-wise.

I'm keeping this open since there are still things that need to be added here.